refactor: new  file getEsBuildTarget
