+++ packages/bundler-webpack/src/utils/browsersList.ts
@@ -1,5 +1,3 @@
-import { DEFAULT_ESBUILD_TARGET_KEYS } from '../constants';
-
 interface IOpts {
   targets: Record<string, any>;
 }
@@ -11,12 +9,4 @@ export function getBrowsersList({ targets }: IOpts) {
       return `${key} >= ${targets[key] === true ? '0' : targets[key]}`;
     })
   );
-}
-
-export function getEsBuildTarget({ targets }: IOpts) {
-  return Object.keys(targets)
-    .filter((key) => DEFAULT_ESBUILD_TARGET_KEYS.includes(key))
-    .map((key) => {
-      return `${key}${targets[key] === true ? '0' : targets[key]}`;
-    });
 }
